+++ docs/CHANGELOG.md
@@ -18,4 +18,6 @@ Changes in libobsd master:
 - Add closefrom()
 - Include io.h in unistd.h on Windows to expose io.h's POSIX functions
 - Add C++ support
+- add timespeccmp(), timespecsub(), and other timeval/timespec macros from
+  sys/time.h